change id type for connection database and fix little but in hardcoded service code

+++ lecture-01/backend/src/main/java/com/example/backend/service/TodoHardcodedService.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Service
 public class TodoHardcodedService {
     private static List<Todo> todos = new ArrayList<>();
-    private static int idCounter = 0;
+    private static long idCounter = 0;
 
     static {
         todos.add(new Todo(++idCounter, "kkh", "learn to Dance", new Date(), false));
